# feat(query): improve like function