--- a/src/query/functions/src/scalars/comparison.rs
+++ b/src/query/functions/src/scalars/comparison.rs
@@ -907,73 +907,41 @@ fn simple_like(
 
 #[test]
 fn test_check_pattern_type() {
-    // OrdinalStr
-    let pattern = "databend".as_bytes();
-    assert_eq!(PatternType::OrdinalStr, check_pattern_type(pattern, false));
-
-    // StartOfPercent
-    let pattern = "%databend".as_bytes();
-    assert_eq!(
-        PatternType::StartOfPercent,
-        check_pattern_type(pattern, false)
-    );
-
-    // EndOfPercent
-    let pattern = "databend%".as_bytes();
-    assert_eq!(
-        PatternType::EndOfPercent,
-        check_pattern_type(pattern, false)
-    );
-
-    // SurroundByPercent
-    let pattern = "%databend%".as_bytes();
-    assert_eq!(
-        PatternType::SurroundByPercent,
-        check_pattern_type(pattern, false)
-    );
-
-    // SimplePattern
     let segments = vec![
         "databend".as_bytes().to_vec(),
         "cloud".as_bytes().to_vec(),
         "data".as_bytes().to_vec(),
         "warehouse".as_bytes().to_vec(),
     ];
-    let pattern = "databend%cloud%data%warehouse".as_bytes();
-    assert_eq!(
-        PatternType::SimplePattern((false, false, segments.clone())),
-        check_pattern_type(pattern, false)
-    );
-    let pattern = "%databend%cloud%data%warehouse".as_bytes();
-    assert_eq!(
-        PatternType::SimplePattern((true, false, segments.clone())),
-        check_pattern_type(pattern, false)
-    );
-    let pattern = "databend%cloud%data%warehouse%".as_bytes();
-    assert_eq!(
-        PatternType::SimplePattern((false, true, segments.clone())),
-        check_pattern_type(pattern, false)
-    );
-    let pattern = "%databend%cloud%data%warehouse%".as_bytes();
-    assert_eq!(
-        PatternType::SimplePattern((true, true, segments)),
-        check_pattern_type(pattern, false)
-    );
-
-    // ComplexPattern
-    let pattern = "databend_cloud%data%warehouse".as_bytes();
-    assert_eq!(
-        PatternType::ComplexPattern,
-        check_pattern_type(pattern, false)
-    );
-    let pattern = "databend\\%cloud%data%warehouse".as_bytes();
-    assert_eq!(
-        PatternType::ComplexPattern,
-        check_pattern_type(pattern, false)
-    );
-    let pattern = "databend%cloud_data%warehouse".as_bytes();
-    assert_eq!(
-        PatternType::ComplexPattern,
-        check_pattern_type(pattern, false)
-    );
+    let test_cases = vec![
+        ("databend", PatternType::OrdinalStr),
+        ("%databend", PatternType::StartOfPercent),
+        ("databend%", PatternType::EndOfPercent),
+        ("%databend%", PatternType::SurroundByPercent),
+        (
+            "databend%cloud%data%warehouse",
+            PatternType::SimplePattern((false, false, segments.clone())),
+        ),
+        (
+            "%databend%cloud%data%warehouse",
+            PatternType::SimplePattern((true, false, segments.clone())),
+        ),
+        (
+            "databend%cloud%data%warehouse%",
+            PatternType::SimplePattern((false, true, segments.clone())),
+        ),
+        (
+            "%databend%cloud%data%warehouse%",
+            PatternType::SimplePattern((true, true, segments)),
+        ),
+        ("databend_cloud%data%warehouse", PatternType::ComplexPattern),
+        (
+            "databend\\%cloud%data%warehouse",
+            PatternType::ComplexPattern,
+        ),
+        ("databend%cloud_data%warehouse", PatternType::ComplexPattern),
+    ];
+    for (pattern, pattern_type) in test_cases {
+        assert_eq!(pattern_type, check_pattern_type(pattern.as_bytes(), false));
+    }
 }