I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

Summary about this PR
Improve like function, when we can't use `OrdinalStr`, `StartOfPercent`, `EndOfPercent` and `SurroundByPercent`, if the like expression only includes `%`, e.g. `'A%r%w'`, we will use `SimplePattern` to get better performance: TPC-H(SF=100) Q13: 8.6s -> 6.3s.

```sql
From TPC-H Q13:
select count(*) from orders where o_comment not like '%pending%deposits%';

TPC-H(SF=100): 4.03s -> 1.61s
```
ci-benchmark:
<img width="1102" alt="截屏2023-07-05 16 17 32" src="https://github.com/datafuselabs/databend/assets/54522439/f57f375d-3937-4df2-a336-c419316d8765">



- Closes #issue