@sundy-li 2023-07-06
Do you have any stateless test cases which cover `SimplePattern ` pattern?