@Dousir9 2023-07-06
> SimplePattern

Let me add some tests.